Fix ineficient InstrusiveSet<> & IntrusiveMap<>.
There were two bugs leading to highly ineficient intrusive set & map:
* The hash function was botched, leading to very unbalanced buckets.
  Now use Fowler-Noll-Vo (FNV) for 32 bits keys.
* The number of buckets was bounded to 512. This was way too small
  when loading netlist above 100K gates. Now put the hard limit to
  2^22, with a clear warning when reached.

For the Black Parrot bench, which has 2.5M gates, we got down from
more than two hours to 5 minutes loading.
  This also solves the problem of slow graphic refresh when nets where
having too much components.

+++ hurricane/src/hurricane/hurricane/SharedName.h
@@ -34,7 +34,7 @@ namespace Hurricane {
     public:
       static void           dump         ();
     public:
-      inline unsigned long  getHash      () const;
+      inline unsigned int   getHash      () const;
              const string&  _getSString  () const { return _string; };
              string         _getTypeName () const { return _TName("SharedName"); };
              string         _getString   () const;
@@ -55,13 +55,13 @@ namespace Hurricane {
 
     private:
       static SharedNameMap* _SHARED_NAME_MAP;
-             unsigned long  _hash;
+             unsigned int   _hash;
              int            _count;
              string         _string;
   };
 
 
-  inline  unsigned long  SharedName::getHash () const { return _hash; }
+  inline  unsigned int  SharedName::getHash () const { return _hash; }
 
 
 } // End of Hurricane namespace.
